Increases the sample frequency from 100/sec to 10,000/sec
For a typical vsync locked application running at 60fps, sampling at just
under twice a frame doesn't seem to give very stable lists of relevent hardware
units because there are a number of units involved that may not be sampled one
second to the next.

This bumps the sample rate to 10,000 instead which is ~ 170 samples per
frame so we tend to hit all the units involved.

It also changes the report threshold to a sample count >= 1, so you don't
see as many units with a percentage of 0.

Signed-off-by: Eric Anholt <eric@anholt.net>