tests: add a test for checking edid reading delays
With base on EDID timing testing, when we take more than 1s to run
xrandr command, something is wrong.. So add this test for testing the time
we take to read the status of all the connectors from sysfs. It should do
us an average picture of how long we'd take to run xrandr (roughtly 2x
that value).

Signed-off-by: Eugeni Dodonov <eugeni.dodonov@intel.com>