config: set-up xorg automatic rebuilding rules
The current code is a noop. Use the same configuration as all
the other xorg modules. This will change in the future but it is less
confusing when all modules behave the same way.
Note that these rules apply to building from a tarball only.

Restore autogen.sh to be identical in all xorg modules.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>