config: the minimum version for autoconf is 2.60
A version later than 2.60 can be used, but no new features from such
a later version can be used in configure.ac.

The toolchain version requirements are documented here:
http://www.x.org/wiki/ModularDevelopersGuide#GNU_Build_System

Minimum version for Automake is 1.10 but we have not written it
in any xorg module so far and there has been no issues.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>