tets/kms_flip: disable correlation check
It's broken since it doesn't take the order the events arrived into
account and so will fall over for longer seq_step values in the vblank
event: Since the flip completes right away, but the vblank later on
they won't have the same timestamp.

For now just give up, we could resurrect this by strictly alternating
between a flip and a vblank wait.

Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>