tests/pc8: add modeset-lpsp and modeset-non-lpsp
The goal of these tests is to exercise the combination of power well
enabling/disabling and PC8 enabling/disabling. The modeset-non-lpsp
test exposes a silent bug on the current code that can just be
detected with the Runtime D3 patches. The Kernel patch that fixes the
bug will add a WARN to unmute it on the PC8-only case.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>