lib/drmtest: Scream harder when igt_exit isn't called for subtest tests
We really need this since otherwise the magic return value handling
for running testcases with piglit (or on QA's validation
infrastructure) doesn't work properly.

We need to be careful though to only install this check on success.

See also the previous commits to sprinkle igt_exit() calls over all
the tests that missed it.

Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>